+++ torch/nn/_intrinsic/qat/modules/__init__.py
@@ -1,8 +1,7 @@
 from __future__ import absolute_import, division, print_function, unicode_literals
 
 from .linear_relu import LinearReLU
-from .conv_relu import ConvReLU2d
-from .conv_bn_relu import ConvBn2d, ConvBnReLU2d
+from .conv_fused import ConvBn2d, ConvBnReLU2d, ConvReLU2d
 
 __all__ = [
     'LinearReLU',
